🧹 Ruff lint fixed stgraph_base.py
